+++ src/tunacode/constants.py
@@ -38,6 +38,10 @@ LOCAL_MAX_LINE_LENGTH = 500
 LOCAL_MAX_FILES_IN_DIR = 20
 DEFAULT_CONTEXT_WINDOW = 200000
 
+# Summary generation thresholds (tokens)
+SUMMARY_THRESHOLD: int = 40_000
+LOCAL_SUMMARY_THRESHOLD: int = 6_000
+
 MAX_CALLBACK_CONTENT = 50_000
 MAX_PANEL_LINES = 20
 MIN_TOOL_PANEL_LINE_WIDTH = 4
