+++ src/tunacode/ui/commands/__init__.py
@@ -2,13 +2,14 @@
 
 from __future__ import annotations
 
-from abc import ABC, abstractmethod
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
     from tunacode.ui.app import TextualReplApp
 
 from tunacode.configuration.settings import ApplicationSettings
+from tunacode.ui.commands.base import Command
+from tunacode.ui.commands.compact import CompactCommand
 from tunacode.ui.styles import STYLE_PRIMARY
 
 # Update command constants
@@ -33,19 +34,6 @@ def _get_package_manager_command(package: str) -> tuple[list[str], str] | None:
         return ([pip_path, "install", "--upgrade", package], "pip")
 
     return None
-
-
-class Command(ABC):
-    """Base class for REPL commands."""
-
-    name: str
-    description: str
-    usage: str = ""
-
-    @abstractmethod
-    async def execute(self, app: TextualReplApp, args: str) -> None:
-        """Execute the command."""
-        pass
 
 
 class HelpCommand(Command):
@@ -541,93 +529,6 @@ class UpdateCommand(Command):
         else:
             app.notify(f"Unknown subcommand: {subcommand}", severity="warning")
             app.notify("Usage: /update [check]")
-
-
-class CompactCommand(Command):
-    name = "compact"
-    description = "Compress conversation history into a summary"
-    usage = "/compact [status]"
-
-    async def execute(self, app: TextualReplApp, args: str) -> None:
-        from tunacode.core.agents.resume import (
-            create_summary_request_message,
-            filter_compacted,
-            generate_summary,
-        )
-        from tunacode.core.agents.resume.summary import SUMMARY_THRESHOLD
-        from tunacode.utils.messaging import estimate_tokens
-
-        session = app.state_manager.session
-        messages = session.messages
-        model = session.current_model
-
-        total_tokens = 0
-        for msg in messages:
-            if isinstance(msg, dict):
-                content = msg.get("content", "")
-                if isinstance(content, str):
-                    total_tokens += estimate_tokens(content, model)
-            elif hasattr(msg, "parts"):
-                for part in msg.parts:
-                    content = getattr(part, "content", None)
-                    if isinstance(content, str):
-                        total_tokens += estimate_tokens(content, model)
-
-        subcommand = args.strip().lower() if args else ""
-
-        if subcommand == "status":
-            app.rich_log.write(f"Messages: {len(messages)}")
-            app.rich_log.write(f"Tokens: ~{total_tokens:,}")
-            app.rich_log.write(f"Threshold: {SUMMARY_THRESHOLD:,}")
-
-            if total_tokens > SUMMARY_THRESHOLD:
-                app.rich_log.write("[yellow]Above threshold - /compact will summarize[/yellow]")
-            else:
-                app.rich_log.write("[dim]Below threshold[/dim]")
-            return
-
-        if len(messages) < 4:
-            app.notify("Not enough messages to compact (need at least 4)", severity="warning")
-            return
-
-        filtered = filter_compacted(messages)
-        if len(filtered) < len(messages) and len(filtered) <= 4:
-            app.notify("Already compacted recently", severity="warning")
-            return
-
-        app.notify("Compacting conversation...")
-
-        try:
-            from tunacode.core.agents import agent_components as ac
-
-            agent = ac.get_or_create_agent(model, app.state_manager)
-            end_index = max(0, len(messages) - 3)
-
-            summary = await generate_summary(
-                agent,
-                messages,
-                model,
-                start_index=0,
-                end_index=end_index,
-            )
-
-            summary_message = create_summary_request_message(summary)
-            messages.append(summary_message)
-            session.update_token_count()
-
-            # Store summary for ctrl+o viewing
-            session.last_summary = summary
-
-            app.rich_log.write(
-                f"[green]Compacted[/green] {end_index} messages "
-                f"[dim](ctrl+o to see summary)[/dim]"
-            )
-            app.rich_log.write(f"[dim]Summary: {summary.token_count} tokens[/dim]")
-            app.notify(f"Compacted! {summary.token_count} token summary")
-            app.state_manager.save_session()
-
-        except Exception as e:
-            app.notify(f"Compact failed: {e}", severity="error")
 
 
 COMMANDS: dict[str, Command] = {
