+++ src/tunacode/ui/commands/base.py
@@ -1,0 +1,22 @@
+"""Base command class for TunaCode REPL."""
+
+from __future__ import annotations
+
+from abc import ABC, abstractmethod
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from tunacode.ui.app import TextualReplApp
+
+
+class Command(ABC):
+    """Base class for REPL commands."""
+
+    name: str
+    description: str
+    usage: str = ""
+
+    @abstractmethod
+    async def execute(self, app: TextualReplApp, args: str) -> None:
+        """Execute the command."""
+        pass
